rename nav.json to metadata.json
